requirements.txt
streamlit
streamlit-lottie
streamlit-webrtc
python-dotenv
langchain-cohere
vosk
sounddevice
pyttsx3
